Organizer localizers into module, add randomized localizer
