+++ docs/src/API/Localizers.md
@@ -1,0 +1,11 @@
+# Localizers
+
+```@meta
+CurrentModule = EnsembleKalmanProcesses.Localizers
+```
+
+```@docs
+Localizer
+RBF
+BernoulliDropout
+```
