+++ src/EnsembleKalmanProcess.jl
@@ -1,5 +1,6 @@
 using ..ParameterDistributions
 using ..DataContainers
+using ..Localizers
 
 using Random
 using Statistics
@@ -14,11 +15,9 @@ export compute_error!
 export update_ensemble!
 export sample_empirical_gaussian, split_indices_by_success
 export SampleSuccGauss, IgnoreFailures, FailureHandler
-export NoLocalization, Delta, RBF, Localizer
 
 abstract type Process end
 #specific Processes and their exports are included after the general definitions
-
 
 # Failure handlers
 abstract type FailureHandlingMethod end
@@ -28,45 +27,6 @@ struct SampleSuccGauss <: FailureHandlingMethod end
 
 struct FailureHandler{P <: Process, FM <: FailureHandlingMethod}
     failsafe_update::Function
-end
-
-# Localizers
-abstract type LocalizationMethod end
-
-struct NoLocalization <: LocalizationMethod end
-struct Delta <: LocalizationMethod end
-
-struct RBF{FT <: Real} <: LocalizationMethod
-    "Length scale defining the RBF kernel"
-    lengthscale::FT
-end
-
-struct Localizer{LM <: LocalizationMethod, T}
-    kernel::Union{AbstractMatrix{T}, UniformScaling{T}}
-end
-
-"Uniform kernel constructor"
-function Localizer(localization::NoLocalization, p::IT, d::IT, T = Float64) where {IT <: Int}
-    kernel = ones(T, p, d)
-    return Localizer{NoLocalization, T}(kernel)
-end
-
-"Delta kernel localizer constructor"
-function Localizer(localization::Delta, p::IT, d::IT, T = Float64) where {IT <: Int}
-    kernel = T(1) * Matrix(I, p, d)
-    return Localizer{Delta, T}(kernel)
-end
-
-"RBF kernel localizer constructor"
-function Localizer(localization::RBF, p::IT, d::IT, T = Float64) where {IT <: Int}
-    l = localization.lengthscale
-    kernel = zeros(T, p, d)
-    for i in 1:p
-        for j in 1:d
-            @inbounds kernel[i, j] = exp(-(i - j) * (i - j) / (2 * l * l))
-        end
-    end
-    return Localizer{RBF, T}(kernel)
 end
 
 ## begin general constructor and function definitions
