+++ test/runtests.jl
@@ -24,7 +24,7 @@ end
         end
     end
 
-    for submodule in ["DataContainers", "ParameterDistributions", "Observations", "EnsembleKalmanProcess"]
+    for submodule in ["DataContainers", "ParameterDistributions", "Observations", "EnsembleKalmanProcess", "Localizers"]
         if all_tests || has_submodule(submodule) || "EnsembleKalmanProcesses" in ARGS
             include_test(submodule)
         end
